% Generated by hand

\name{bike_gears0}
\alias{bike_gears0}
\title{Gear Count Zero}
\description{
This function computes the gear count.
}
\usage{
  r <- bike_gears0(3L, 7L)
}
\arguments{
  \item{nFront}{count of front chainrings}
  \item{nBack}{count of rear cassette rings}
}
\author{Maintainer: \email{R@gams.com}}

\examples{
  r <- (3L, 5L);
}
\keyword{ biking }
