% Generated by hand

\name{bike_gears1}
\alias{bike_gears1}
\title{Gear Count One}
\description{
This function computes the gear count.
}
\usage{
  r <- bike_gears1(3L, 7L)
}
\arguments{
  \item{nFront}{count of front chainrings}
  \item{nBack}{count of rear cassette rings}
}
\author{Maintainer: \email{R@gams.com}}

\examples{
  r <- bike_gears1(3L, 5L);
}
\keyword{ biking }
