% Generated by hand

\name{biking_function}
\alias{biking_function}
\title{A Biking Function}
\description{
This function allows you to express your love of biking.
}
\usage{
  biking_function(love = TRUE)
}
\arguments{
  \item{love}{Do you love biking? Defaults to TRUE.}
}
\author{Original coding by Rishabh Jain.  Adopted and packaged by
Steve Dirkse.  Maintainer: \email{R@gams.com}}

\examples{
  biking_function();
  \dontrun{
    biking_function(0);
    biking_function(T);
  }
}
\keyword{ biking }
